Implemented visualization client as website
A tornado server implemented in Python serves audio/spectral data to
some Javascript in the browser. Server and client communicate over a
websocket. A minimal protocol has been established with which client
and server can communicate JSON data and binary data.

- All messages have a JSON part that contains the key 'type', for the
type of the message.
- Regular messages contain just the JSON data structure.
- Binary messages contain an int32 with the size of the JSON data
structure in bytes, aligned to 8 byte. Then, binary data as specified
in the header.