transpose texture and fix shaders
The spectrogram data sent from Python is blocks x freqs
(column-major), in order to be able to select a set of blocks as a
contiguous chunk of memory. However, the texture needs to be
constructed freqs x blocks (row-major) in order for GL to render it
correctly. I used to solve this by rotating the texture coordinates.
This however makes explicit texture indexing complicated, since x and
y are switched. Now, I transpose the data in Javascript before
uploading it to the GL. This takes about 100 ms for 3 min of audio.